# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-averageweighted",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the weighted average of `values` with `weights`.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits float dependencies:\n\n1. `sumwv = fold floatadd over floatmultiply(values[i], weights[i])` (start `0`);\n2. `sumw  = fold floatadd over weights` (start `0`);\n3. `result = floatdivide(sumwv, sumw)`.\n\n`values` and `weights` must be non-empty and of equal length (`422`\notherwise). If a dependency is unreachable it reports itself degraded\n(`503`); if a dependency rejects an operand it forwards the `422`.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightedAverageResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid input, or a dependency rejected an operand (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "values",
          "weights"
        ],
        "properties": {
          "values": {
            "type": "object",
            "description": "The values being averaged."
          },
          "weights": {
            "type": "object",
            "description": "The weight applied to each corresponding value."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "WeightedAverageResponse": {
        "type": "object",
        "required": [
          "values",
          "weights",
          "result"
        ],
        "properties": {
          "result": {
            "type": "number",
            "format": "double"
          },
          "values": {
            "type": "object"
          },
          "weights": {
            "type": "object"
          }
        }
      }
    }
  }
}